Change the API to parsing number a bit.
Since we are marshalling the entire string into the C world, it is
not a good idea to give this function the whole input (e.g., the rest
of the file).   This changes the API so that we don't return the rest
of the input, which makes it more obvious that we expect to only be
given the string corresponding to the number.