Don't leak the internal `LimbT` type in the public API.
Uniformly use the architecture-dependent `Word` type to
represent bits of precision, and the `Int` type to represent
exponent bits.  Yes, it's odd to use a signed type for one
and an unsigned type for the other, but that's what the types
`libbf.h` are.

Hopefull this should allow the library to work properly on
32-bit systems.