Update vendored libbf to most recent version used in quickjs
The `libbf` version that was previously vendored (`2020-01-19`) simply did not
work properly on non-x86 64-bit architectures, as observed in #22.
Unfortunately, that is the most recent official `libbf` release, but there is
also another, more up-to-date version of `libbf` (by the same author) that is
bundled with the `quickjs` project.  This version of `libbf` (bundled as part
of the `2021-03-27` `quickjs` release, see
https://github.com/bellard/quickjs/commit/b5e62895c619d4ffc75c9d822c8d85f1ece77e5b)
_does_ work properly on other 64-bit architectures.

This commit updates the version of `libbf` that we vendor to the more
up-to-date `quickjs` version, thereby fixing #22 in the process.